Unable to open PR editor: Cannot read properties of undefined (reading 'toLowerCase' Fixes #6385
